Refactored transaction modal
